// client-monorepo/libs/applets/wealth/src/lib/components/core/models/instruments.enum.ts
export enum EInstruments {
  SINA = 'SINA',
  GOLRANG = 'GOLRANG',
  BEHGOZIN = 'BEHGOZIN',
  DAMAVAND = 'DAMAVAND',
  GARDESHGARI = 'GARDESHGARI',
  LOTOS = 'LOTOS',
  REGULATIONS = 'REGULATIONS',
  NEGIN = 'NEGIN',
  GANJ = 'GANJ',
  RISK = 'RISK',
  IPO = 'IPO',
  KARDAN = 'KARDAN',
  AFAGH = 'AFAGH',
  ACORD = 'ACORD',
  AKAM = 'AKAM',
  VISTA = 'VISTA',
  ZARVAN = 'ZARVAN',
  KAMYAB = 'KAMYAB',
  VBAZAR = 'VBAZAR',
  TREASURY = 'TREASURY',
}

export enum EpdfType {
  PROSPECTUS = 'PROSPECTUS',
  STATUTE = 'STATUTE',
  CONTRACT = 'CONTRACT',
  REGULATIONS = 'REGULARATION',
  RISKSTATEMEBT = 'RISKSTATEMEBT',
}